Update BaseGradeBook's GetGPA Method
BaseGradeBook's GetGPA method returns an int based on the student's letter grade (their GPA or Grade Point Average). When a grade book's IsWeighted property is true and the student type is Honors or DualEnrolled, add 1 to the GPA before returning it.

An unweighted A is worth 4 points.
A weighted A is worth 5 points.
An unweighted B is worth 3 points.
A weighted B is worth 4 points and so on.